% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checknzv.R
\name{checknzv}
\alias{checknzv}
\title{Check nzv}
\usage{
checknzv(df, the_nzv = NULL, n = min(10, round(ncol(df)/10)),
  v_label = f_train_label, verbose = T, ...)
}
\arguments{
\item{df}{the dataset as a data.frame or matrix}

\item{the_nzv}{default NULL, the already calculated nzv by \code{caret::nzv}}

\item{n}{first n features with high nzv}

\item{v_label}{vector of lable}

\item{verbose}{logical}

\item{...}{other parameters}
}
\value{
a list of features with high nzv checks
}
\description{
This function is needed also because I would like to have the insight of what
nzv can contribute to feature selection / filtering. Some interesting topics:
do we always want to drop zv? when do we want to drop some nzv also?
}

