% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df-manipulation.R
\name{convertAllToCharacter}
\alias{convertAllToCharacter}
\title{Convert all columns to character type.}
\usage{
convertAllToCharacter(df)
}
\arguments{
\item{df}{a data.frame}
}
\value{
a character data.frame
}
\description{
Convert all columns in a data.frame to character type. Simply does not make
sense to convert to other types, and good practice to keep the original types
somewhere.
}
\author{
Siyu Sun
}

