% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVfeatures_functions.R
\name{cosim}
\alias{cosim}
\title{Complex/simple events counts
Takes as input the path to the clusters (Linx output) and the samples
and calculates the number of complex events and simple events per sample}
\usage{
cosim(samples)
}
\description{
Complex/simple events counts
Takes as input the path to the clusters (Linx output) and the samples
and calculates the number of complex events and simple events per sample
}
