% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createfeatures.R
\name{createBinaryFeatures}
\alias{createBinaryFeatures}
\title{Create Statistical Summary Features}
\usage{
createBinaryFeatures(dt, cnames = NULL, operator = "-", verbose = T)
}
\description{
create row-aggregated features rowsum, mean, median, mode, sd, max, min,
time: first last,
counts: >0, nonNA,
}

