% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createfeatures.R
\name{createDiffFeatures}
\alias{createDiffFeatures}
\title{Create Statistical Summary Features}
\usage{
createDiffFeatures(dt, cnames = NULL, verbose = T)
}
\description{
create row-aggregated features rowsum, mean, median, mode, sd, max, min,
time: first last,
counts: >0, nonNA,
}

