% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df-manipulation.R
\name{fillColbyName}
\alias{fillColbyName}
\title{Fill a data.frame by give column names}
\usage{
fillColbyName(df, fillnames)
}
\arguments{
\item{df}{a data.frame}

\item{fillnames}{a character vector of kept column names}
}
\value{
a data.frame
}
\description{
Bind columns of NA_character_ to data.frame that are not matched with given
desirable column names. Tip: always cut first before fill.
}
\author{
Siyu Sun
}

