% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-helpers.R
\name{findAllUniqueValues}
\alias{findAllUniqueValues}
\title{Find all unique values within each column of one data.frame}
\usage{
findAllUniqueValues(l, n = 1, onlyunique = T)
}
\arguments{
\item{l}{a list}
}
\value{
a character vector
}
\description{
Find all unique values within each column of one data.frame. The union of all
values will be returned as a character vector. Note: only supply a list with
list elements of same class, otherwise NULL will be returned.
}
\seealso{
Other list helpers: \code{\link{findAllVarNames}},
  \code{\link{getVarNameFreq}}
}

