% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checknzv.R
\name{getcrosstabledf}
\alias{getcrosstabledf}
\title{Get Cross Table as customized data.frame}
\usage{
getcrosstabledf(df, feature, v_label, breaks = 20, useNA = "ifany")
}
\arguments{
\item{v_label}{a lable vector}

\item{breaks}{bar to avoid features with too many unique values, if exceed
the bar, \code{cut} the feature to n==breaks categories.}

\item{v_feature}{a feature vector}
}
\value{
a data.frame
}
\description{
When training model or examining features, one trick is to check the cross
table of one feature and the label column. Too lazy to do this every time.
}

