% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feat-date.R
\name{getDateFeatures}
\alias{getDateFeatures}
\title{Get Date Features from a vector of dates}
\usage{
getDateFeatures(df, colname, tz = "UTC", cutdate = Sys.Date())
}
\arguments{
\item{df}{a data.frame}

\item{colname}{the column name of the date column}

\item{curr_date}{default, \code{Sys.Date()}}
}
\value{
a data.frame with expanded feature columns
}
\description{
used lubridate functions.
}

