% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feat-date.R
\name{getDateOffset}
\alias{getDateOffset}
\title{Get Max. Date Offset}
\usage{
getDateOffset(df, date_colname, units = "years", format = "\%d\%m\%Y",
  curr_date = Sys.Date())
}
\arguments{
\item{df}{a data.frame}

\item{date_colname}{the column name of the date column}

\item{units}{expected units, "years" or all other units accepted by \code{difftime}}

\item{curr_date}{default, \code{Sys.Date()}}
}
\value{
a numeric value
}
\description{
Calculate the maximum date offset between a date column in a data.frame and
the current date.
}

