% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checknzv.R
\name{getnzvasdf}
\alias{getnzvasdf}
\title{Get nzv as a nice formatted data.frame}
\usage{
getnzvasdf(df, the_nzv = NULL, onlyzv = F, onlynzv = T, verbose = T)
}
\arguments{
\item{df}{the dataset as a data.frame or matrix}

\item{the_nzv}{default NULL, the already calculated nzv by \code{caret::nzv}}

\item{onlyzv}{logical, if TRUE, only zero variance will be returned}

\item{onlynzv}{logical, if TRUE, only nzv but not zv will be returned}

\item{verbose}{logical}
}
\value{
a data.frame of customized nzv
}
\description{
This function is needed simply because I would like my own nzv and am too
lazy to type the same code every time.
}

