% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-helpers.R
\name{getVarNameFreq}
\alias{getVarNameFreq}
\title{Calculate variable name frequency with a list of list elements.}
\usage{
getVarNameFreq(l)
}
\arguments{
\item{l}{a list}
}
\value{
a data.frame with two columns: varnames and Freq
}
\description{
Find out which variable names are most shared by all list elements, which are
not. If a variable name is with very low frequency, it should probably be
regarded (but check for exceptions).
}
\seealso{
Other list helpers: \code{\link{findAllUniqueValues}},
  \code{\link{findAllVarNames}}
}

