% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVfeatures_functions.R
\name{is.chromotripsis}
\alias{is.chromotripsis}
\title{Function for identification of the presence of chromotripsis in each sample
Using Shatterseek for identifying the regions eligible for chromotripsis and their criteria
for the classification of the events as "high-confidence" or "low-confidence"
If there is at least one high-confidence region in the sample the esit is 1 otherwise 0}
\usage{
is.chromotripsis(samples, segmentation)
}
\description{
Function for identification of the presence of chromotripsis in each sample
Using Shatterseek for identifying the regions eligible for chromotripsis and their criteria
for the classification of the events as "high-confidence" or "low-confidence"
If there is at least one high-confidence region in the sample the esit is 1 otherwise 0
}
