% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVfeatures_functions.R
\name{logR_quantiles}
\alias{logR_quantiles}
\title{20 and 80 quantiles of loR
Takes as input the sample list and the segmentation file (purple output)
it inner_joins the tables and gives back the samples table added with the two quantiles}
\usage{
logR_quantiles(samples, segmentation)
}
\description{
20 and 80 quantiles of loR
Takes as input the sample list and the segmentation file (purple output)
it inner_joins the tables and gives back the samples table added with the two quantiles
}
