% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printchecks.R
\name{printFeaturePreChecks}
\alias{printFeaturePreChecks}
\title{Prints accelerate raw feature dataset pre-checks}
\usage{
printFeaturePreChecks(l)
}
\arguments{
\item{l}{a list}
}
\value{
invisible NULL
}
\description{
Refer to Cookbook Feature for this print out workflow. Note: currently the
strict rule is that only a list can be accepted as input.
}
\examples{
\dontrun{
printFeaturePreChecks(lmxburo)
}
}
\author{
Siyu Sun
}
\seealso{
Other print.checks: \code{\link{printOneFeatureChecks}}
}

