% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printchecks.R
\name{printOneFeatureChecks}
\alias{printOneFeatureChecks}
\title{Prints accelerate pre-checks of a single feature}
\usage{
printOneFeatureChecks(l)
}
\arguments{
\item{l}{a list, contains}
}
\value{
invisible NULL
}
\description{
Refer to Cookbook Feature for this print out workflow. Note: currently only
for lists with same data.frame variables/columns
}
\examples{
\dontrun{
printOneFeatureChecks(lmxburo)
}
}
\author{
Siyu Sun
}
\seealso{
Other print.checks: \code{\link{printFeaturePreChecks}}
}

