% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df-manipulation.R
\name{renameOneCol}
\alias{renameOneCol}
\title{Rename one column to another column's name.}
\usage{
renameOneCol(df, old, new)
}
\arguments{
\item{df}{a data.frame}

\item{old}{a character vector of old name}

\item{new}{a character vector of new name}
}
\value{
a data.frame
}
\description{
This function is needed when there is an obvious colname error/mispell with
some observations. This observation does not have a variable name like most
other observations, but has an abnormal one that can be confirmed to be the
same as another variable.
}
\author{
Siyu Sun
}

