% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVfeatures_functions.R
\name{trater}
\alias{trater}
\title{Intrachromosomal and interchromosomal SVs counts
Takes as input the samples, extract the pairs object from Purple svs output
and calculates the number of intra/inter-chromosomal events per sample
according to the pairs of SVs laying on the same chromosome or not}
\usage{
trater(samples)
}
\description{
Intrachromosomal and interchromosomal SVs counts
Takes as input the samples, extract the pairs object from Purple svs output
and calculates the number of intra/inter-chromosomal events per sample
according to the pairs of SVs laying on the same chromosome or not
}
