% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vgrepl.R
\name{vgrepl}
\alias{vgrepl}
\title{vectorize grepl}
\usage{
vgrepl(vpattern, vx, operator = "or")
}
\arguments{
\item{vpattern}{vector of patterns}

\item{vx}{vector of x}

\item{operator}{either "or" or "and"}
}
\value{
logical vector of same length of vx
}
\description{
vectorize grepl
}

